% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASRgenomics-package.R
\docType{package}
\name{ASRgenomics-package}
\alias{ASRgenomics}
\alias{ASRgenomics-package}
\title{ASRgenomics: Complementary Genomic Functions}
\description{
Presents a series of molecular and genetic routines in the R environment with the aim of assisting in analytical pipelines before and after the use of 'asreml' or another library to perform analyses such as Genomic Selection or Genome-Wide Association Analyses. Methods and examples are described in Gezan, Oliveira, Galli, and Murray (2022) \url{https://asreml.kb.vsni.co.uk/wp-content/uploads/sites/3/ASRgenomics_Manual.pdf}.
}
\author{
\strong{Maintainer}: Salvador Gezan \email{salvador.gezan@vsni.co.uk}

Authors:
\itemize{
  \item Darren Murray \email{darren.murray@vsni.co.uk}
  \item Amanda Avelar de Oliveira \email{support@vsni.co.uk}
  \item Giovanni Galli \email{giovanni.galli@vsni.co.uk}
}

Other contributors:
\itemize{
  \item VSN International [copyright holder]
}

}
\keyword{internal}
