% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{callrate}
\alias{callrate}
\title{Estimates call rate}
\usage{
callrate(M = NULL, margin = c("row", "col"))
}
\arguments{
\item{M}{The additive \eqn{n \times p} matrix with any coding (default = \code{NULL}).}

\item{margin}{A character indicating the margin for call rate calculations.
Options are: \code{row} and \code{col} (default = \code{row}).}
}
\value{
A vector containing the call rate.
}
\description{
Estimates call rate
}
\keyword{internal}
