% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{check.data.mode_}
\alias{check.data.mode_}
\title{Check data mode}
\usage{
check.data.mode_(data_ = NULL, mode_ = NULL)
}
\arguments{
\item{data_}{Data to be checked.}

\item{mode_}{The expected mode of data_.}
}
\description{
Check data mode
}
\keyword{internal}
