% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{check.data_}
\alias{check.data_}
\title{Check data class}
\usage{
check.data_(data_ = NULL, class_ = NULL)
}
\arguments{
\item{data_}{Data to be checked.}

\item{class_}{The expected class of data_.}
}
\description{
Check data class
}
\keyword{internal}
