% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{check.logical_}
\alias{check.logical_}
\title{Check logical arguments}
\usage{
check.logical_(arg_ = NULL)
}
\arguments{
\item{arg_}{The boolean argument to be checked.}
}
\description{
Check logical arguments
}
\keyword{internal}
