% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kinship_heatmap.R
\name{.heat_kinship}
\alias{.heat_kinship}
\title{Kinship heatmaps (adapted from superheat)}
\usage{
.heat_kinship(
  K,
  dendrogram = TRUE,
  clustering.method = c("hierarchical", "kmeans"),
  dist.method = c("euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski"),
  row.label = TRUE,
  col.label = FALSE
)
}
\arguments{
\item{K}{A numeric matrix.}

\item{dendrogram}{Logical. Draw a column dendrogram.}

\item{clustering.method}{Clustering method. Supports "hierarchical" and "kmeans".}

\item{dist.method}{Distance method passed to dist().}

\item{row.label}{Logical. Draw row labels.}

\item{col.label}{Logical. Draw column labels.}
}
\value{
Invisibly returns a list, with plot,
membership.cols, membership.rows, order.rows, order.cols, and heat.pal.values.
}
\description{
Implements a subset of superheat behavior used by kinship.heatmap().
}
\details{
Credits to the superheat package by Barter and Yu (2020).
}
\keyword{internal}
