% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{dummy.map_}
\alias{dummy.map_}
\title{Creates a dummy map if not provided}
\usage{
dummy.map_(marker.id = NULL, message = TRUE)
}
\arguments{
\item{marker.id}{vector with names of markers to compose dummy map.}

\item{message}{logical value indicating whether diagnostic messages should be printed on screen (default = \code{TRUE}).}
}
\value{
Data frame with dummy map. A single chromosome/linkage group is created and marker
distances are a sequence from one to the number of markers.
}
\description{
Creates a dummy map if not provided
}
\keyword{internal}
