% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{heterozygosity}
\alias{heterozygosity}
\title{Estimates observed and expected heterozygosity}
\usage{
heterozygosity(M = NULL)
}
\arguments{
\item{M}{The additive \eqn{n \times p} matrix coded with 0, 1, 2 coding (default = \code{NULL}).}
}
\value{
A list with vectors containing the observed (ho) and expected (he) heterozygosity.
}
\description{
Estimates observed and expected heterozygosity
}
\keyword{internal}
