% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{Kinv.condition}
\alias{Kinv.condition}
\title{Assess condition of the inverse of \strong{K}}
\usage{
Kinv.condition(Kinv)
}
\arguments{
\item{Kinv}{An inverse relationship matrix in full or sparse form.}
}
\value{
An object of class character with the condition of the matrix:
well-conditioned or ill-conditioned.
}
\description{
Assess condition of the inverse of \strong{K}
}
\keyword{internal}
