% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{maf}
\alias{maf}
\title{Estimates minor allele frequency (MAF)}
\usage{
maf(M = NULL)
}
\arguments{
\item{M}{The additive \eqn{n \times p} matrix coded with 0, 1, 2, or \code{NA}
(default = \code{NULL}).}
}
\value{
A vector with the MAF of each molecular marker
}
\description{
Estimates minor allele frequency (MAF)
}
\keyword{internal}
