% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{silent_}
\alias{silent_}
\title{Function that silences everything (e.g., \code{cat()}, \code{print()}, \code{message()}, ...)}
\usage{
silent_(code)
}
\arguments{
\item{code}{code to be silenced.}
}
\value{
None.
}
\description{
Function that silences everything (e.g., \code{cat()}, \code{print()}, \code{message()}, ...)
}
\keyword{internal}
